Separate lyric time-line into individual class.
